+++ src/shared/components/palette/case-flag/read-case-flag-field.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { CaseTab } from '../../../domain';
+import { CaseField, CaseTab } from '../../../domain';
 import { FieldsUtils } from '../../../services/fields';
 import { PaletteContext } from '../base-field/palette-context.enum';
 import { AbstractFieldReadComponent } from '../base-field/abstract-field-read.component';
@@ -15,7 +15,8 @@ export class ReadCaseFlagFieldComponent extends AbstractFieldReadComponent imple
 
   public flagsData: Flags[];
   public partyLevelCaseFlagData: Flags[];
-  public caseLevelCaseFlagData: Flags[];
+  public caseLevelCaseFlagData: Flags;
+  public readonly caseLevelCaseFlagsFieldId = 'caseFlags';
 
   constructor(
     private readonly route: ActivatedRoute
@@ -35,7 +36,7 @@ export class ReadCaseFlagFieldComponent extends AbstractFieldReadComponent imple
           .some(caseField => caseField.field_type.type === 'FlagLauncher'))
         [0].fields.reduce((flags, caseField) => {
           if (FieldsUtils.isFlagsCaseField(caseField) && caseField.value) {
-            flags.push(this.mapValueToFlagsObject(caseField.value));
+            flags.push(this.mapCaseFieldToFlagsObject(caseField));
           }
           return flags;
         }, []) as Flags[];
@@ -46,25 +47,28 @@ export class ReadCaseFlagFieldComponent extends AbstractFieldReadComponent imple
       this.flagsData = [];
       Object.keys(this.formGroup.controls).filter(
         controlName => FieldsUtils.isFlagsCaseField(this.formGroup.controls[controlName]['caseField']))
-          .forEach(key => this.flagsData.push(this.mapValueToFlagsObject(this.formGroup.controls[key]['caseField'].value)));
+          .forEach(key => {
+            if (this.formGroup.controls[key]['caseField'].value) {
+              this.flagsData.push(this.mapCaseFieldToFlagsObject(this.formGroup.controls[key]['caseField']));
+            }
+          });
     }
 
-    // TODO: In future, this needs to separate party-level and case-level flags
-    this.generateCaseFlagData();
+    // Separate the party-level and case-level flags
+    this.partyLevelCaseFlagData = this.flagsData.filter(
+      flagsInstance => flagsInstance.flagsCaseFieldId !== this.caseLevelCaseFlagsFieldId);
+    // There will be only one case-level flags instance containing all case-level flag details
+    this.caseLevelCaseFlagData = this.flagsData.find(
+      flagsInstance => flagsInstance.flagsCaseFieldId === this.caseLevelCaseFlagsFieldId);
   }
 
-  public generateCaseFlagData(): void {
-    // Temporary assignment of all flags data to party level until it is known how party-level and case-level will be
-    // distinguished
-    this.partyLevelCaseFlagData = this.flagsData;
-  }
-
-  private mapValueToFlagsObject(value: object): Flags {
+  private mapCaseFieldToFlagsObject(caseField: CaseField): Flags {
     return {
-      partyName: value['partyName'],
-      roleOnCase: value['roleOnCase'],
-      details: value['details']
-        ? ((value['details']) as any[]).map(detail => {
+      flagsCaseFieldId: caseField.id,
+      partyName: caseField.value['partyName'],
+      roleOnCase: caseField.value['roleOnCase'],
+      details: caseField.value['details'] && caseField.value['details'].length > 0
+        ? ((caseField.value['details']) as any[]).map(detail => {
           return Object.assign({}, ...Object.keys(detail.value).map(k => {
             switch (k) {
               // These two fields are date-time fields